Add support for WeirdObjects, and fix compilation.
Basic support for WeirdObjects is added. Int literals and String
literals are converted to string objects, however scopifying is not
currently executed so they are never freed.

Furthermore, returning anything from main() is also not supported, but
will be soon, since we just need to convert WeirdIntegers to int
primitives.

Plus, compilation was broken on machines whose compilers already
supplied strdup. This has been fixed with a compile-time check for the
presence of strdup.